+++ examples/react-tldraw/src/App.tsx
@@ -17,7 +17,7 @@ Amazon AWS S3.
 export default function App() {
   const fileSystemEvents = useFileSystem();
   const { ...events } = useMultiplayerState(
-    `tldraw-${(new Date()).toISOString().substring(0, 10).replace(/-/g, '')}`
+    `tldraw-${new Date().toISOString().substring(0, 10).replace(/-/g, '')}`,
   );
   const component = { Cursor: CustomCursor };